pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hermes-aegis"
version = "0.1.0"
description = "Security hardening layer for Hermes Agent"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "cryptography>=41.0",
    "keyring>=24.0",
    "mitmproxy>=10.0",
    "pyyaml>=6.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
container = ["docker>=7.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pyyaml>=6.0",
    "responses>=0.26.0",
]
demo = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
]

[project.scripts]
hermes-aegis = "hermes_aegis.cli:main"
